requirements.txt
redis_cache_decorator
cached_property
xxhash

exifyay
git+http://github.com/smarnach/pyexiftool.git#egg=pyexiftool
git+http://github.com/jed-frey/get_files.git#egg=get_files

versioneer
rq
PyMySQL

black

# Testing
pytest
pytest-html

cython

Pillow

# Development 
jupyter
notebook

## Host Requirements
#16.04 18.04: python3
#16.04 18.04: python3-venv
#16.04 18.04: python3-dev
